// App/Literals/Enums.ts
export enum SaveStatusEnum {
  Failure = 'Failure',
  None = 'None',
  Success = 'Success',
}

export enum ScreenNameEnum {
  App = 'App',
  Home = 'Home',
  Launcher = 'Launcher',
  List = 'List',
  Login = 'Login',
  SignUp = 'SignUp',
}

export enum FileTypeEnum {
  Documents = 'documents',
  Images = 'images',
}